refactor: rename 'amount' formatter as 'tokens' util
