+++ src/features/stake/Stake.tsx
@@ -1,11 +1,11 @@
 import { useStaking } from 'src/features/stake/useStaking';
 import { SwapForm } from 'src/features/swap/SwapForm';
 import { Switcher } from 'src/features/swap/Switcher';
-import { toCeloWei } from 'src/formatters/amount';
 import { CenteredLayout } from 'src/layout/CenteredLayout';
 import { useAccountContext } from 'src/providers/AccountProvider';
 import toast from 'src/services/toast';
 import { Celo } from 'src/types/units';
+import { toCeloWei } from 'src/utils/tokens';
 
 export const Stake = () => {
   const { stake, celoExchangeRate, estimateDepositValue, estimateStakingFee } = useStaking();
