+++ src/features/swap/BalanceTools.tsx
@@ -1,6 +1,6 @@
 import { DISPLAY_DECIMALS } from 'src/config/consts';
-import { fromCeloWei, fromStCeloWei } from 'src/formatters/amount';
 import { CeloWei, StCeloWei } from 'src/types/units';
+import { fromCeloWei, fromStCeloWei } from 'src/utils/tokens';
 
 interface BalanceToolsProps {
   balance: CeloWei | StCeloWei;
