+++ src/features/swap/SwapForm.tsx
@@ -7,9 +7,9 @@ import { INPUT_DECIMALS } from 'src/config/consts';
 import { CostsSummary } from 'src/features/swap/CostsSummary';
 import { ReceiveSummary } from 'src/features/swap/ReceiveSummary';
 import { TokenCard } from 'src/features/swap/TokenCard';
-import { fromWeiRounded } from 'src/formatters/amount';
 import { useExchangeContext } from 'src/providers/ExchangeProvider';
 import { CeloWei, StCeloWei } from 'src/types/units';
+import { fromWeiRounded } from 'src/utils/tokens';
 import { BalanceTools } from './BalanceTools';
 import { SubmitButton } from './SubmitButton';
 import { StakeToken } from './types';
