+++ src/features/unstake/Unstake.tsx
@@ -1,11 +1,11 @@
 import { SwapForm } from 'src/features/swap/SwapForm';
 import { Switcher } from 'src/features/swap/Switcher';
 import { useUnstaking } from 'src/features/unstake/useUnstaking';
-import { toStCeloWei } from 'src/formatters/amount';
 import { CenteredLayout } from 'src/layout/CenteredLayout';
 import { useAccountContext } from 'src/providers/AccountProvider';
 import toast from 'src/services/toast';
 import { StCelo } from 'src/types/units';
+import { toStCeloWei } from 'src/utils/tokens';
 import { PendingWithdrawal } from './PendingWithdrawal';
 
 export const Unstake = () => {
