+++ src/features/unstake/useUnstaking.ts
@@ -1,12 +1,12 @@
 import BigNumber from 'bignumber.js';
 import { useCallback } from 'react';
 import { GAS_LIMIT, GAS_PRICE } from 'src/config/consts';
-import { fromStCeloWei, toStCeloWei } from 'src/formatters/amount';
 import { useContracts } from 'src/hooks/useContracts';
 import { useAccountContext } from 'src/providers/AccountProvider';
 import { useExchangeContext } from 'src/providers/ExchangeProvider';
 import api from 'src/services/api';
 import { StCelo, StCeloWei } from 'src/types/units';
+import { fromStCeloWei, toStCeloWei } from 'src/utils/tokens';
 
 export function useUnstaking() {
   const { address, loadBalances, loadPendingWithdrawals } = useAccountContext();
